requirements.txt
wheel==0.37.1
asn1crypto==0.24.0
click==8.0.4
Flask==2.0.3
idna==2.6
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==2.0.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
requests==2.25.1
kafka-python==2.0.2
pytest==7.1.3
pandas==1.5.1
tqdm==4.64.1
torch==1.12.1
pyyaml==5.4.1
colorlog==6.7.0
tensorboard==2.10.1
colorama==0.4.4
sklearn
coverage==6.5.0
